Move compiler into `syntactic_environment` from `machine`
